requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Redis & Celery
redis==5.0.4
celery==5.3.6

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Parsing
beautifulsoup4==4.12.3
lxml==5.1.0
httpx==0.27.0
python-magic==0.4.27  # для определения MIME-типов

# Logging
python-json-logger==2.0.7

# Utils
python-dateutil==2.9.0
email-validator==2.1.1